get rid of circular import
