+++ dashboard/utils_test.py
@@ -7,6 +7,8 @@ from unittest.mock import (
     MagicMock,
     PropertyMock
 )
+
+from django.test import override_settings
 from django.urls import reverse
 
 import pytz
@@ -732,6 +734,21 @@ class MMTrackTest(MockedESTestCase):
             )
             FinalGradeFactory.create(user=self.user, course_run=course_run, grade=grade, passed=True)
         assert mmtrack.get_best_final_grade_for_course(finaid_course).grade == 0.8
+
+    @override_settings(FEATURES={"USE_COMBINED_FINAL_GRADE": True})
+    def test_get_overall_final_grade_for_course(self):
+        """
+        Test for get_overall_final_grade_for_course to return CombinedFinalGrade for course
+        """
+        mmtrack = MMTrack(
+            user=self.user,
+            program=self.program_financial_aid,
+            edx_user_data=self.cached_edx_user_data
+        )
+        finaid_course = self.crun_fa.course
+        assert mmtrack.get_overall_final_grade_for_course(finaid_course) == ""
+        CombinedFinalGrade.objects.create(user=self.user, course=finaid_course, grade="74")
+        assert mmtrack.get_overall_final_grade_for_course(finaid_course) == "74"
 
     def test_get_best_proctored_exam_grade(self):
         """
