+++ grades/constants.py
@@ -2,6 +2,9 @@
 Constants for the grades app
 """
 
+COURSE_GRADE_WEIGHT = 0.4
+EXAM_GRADE_WEIGHT = 0.6
+
 
 class FinalGradeStatus:
     """
